fix some warning

+++ ArdaCore/Core.h
@@ -8,8 +8,8 @@ using int8 = signed char;
 using uint8 = unsigned char;
 using int16 = signed short;
 using uint16 = unsigned short;
-using int32 = signed long;
-using uint32 = unsigned long;
+using int32 = signed int;
+using uint32 = unsigned int;
 using int64 = signed long long;
 using uint64 = unsigned long long;
